添加canvas测试工程代码

+++ CanvasContext/geometry/FloatPoint.cpp
@@ -88,7 +88,7 @@ FloatPoint::operator SkPoint() const
 
 FloatPoint FloatPoint::narrowPrecision(double x, double y)
 {
-    return FloatPoint((float)(x), (float)(y));
+    return FloatPoint((x), (y));
 }
 
 float findSlope(const FloatPoint& p1, const FloatPoint& p2, float& c)
